requirements.txt
numpy
opencv-python
torch
torchvision
scikit-learn
matplotlib
pandas
PyYAML
tqdm
Pillow

